+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# HEAD
+* Supports scope in config to define whether a user can join an experiment or not.
+
 # 0.2.0 (2017-08-15)
 * **API change**: If all rules failed, `ab_test` returns nil, instead of the first variant.